extend the tags table to include an isDefn field
This patch extends the sqlite database with a new
field 'isDefn', which is populated by calling
the libclang API function

   unsigned clang_isCursorDefinition (CXCursor)

A potential issue with this patch is that the database
is not versioned, and if the new server is run against
an old database, things silently fail.  I don't know
if a return code from sqlite is not being checked
or exactly what the issue might be.

Ideally an API version number would be stored in the
database and if the server detects an old database,
a new column would be added (and perhaps all the tags
entries removed).  Unfortunately I don't know enough sqlite
to make this happen.  But if you want to pursue it,
I can ask for help.

I also don't know how to extend the internal unit tests
to test for this case.  But 'make test' continues to pass,
and I have eyeballed a database for my application and the
results look good!